# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "One Medy API Documentation",
        "description": "One Medy APi Service to be consumed by client Application",
        "contact": {
            "email": "info@walexbiznig.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/administrator/summary": {
            "get": {
                "tags": [
                    "Administrator"
                ],
                "summary": "To get count of summary of the application",
                "description": "Returns counts",
                "operationId": "getSummary",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Size"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/administrator/appointments": {
            "get": {
                "tags": [
                    "Administrator"
                ],
                "summary": "To get all appointments details",
                "description": "Returns Appointments",
                "operationId": "getAppointments",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Size"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}